Move random classes to zrandomcodes package - To improve package structure and maintainability
