+++ src/test/java/com/engineerLiberty/techieAssessment/TimeInWordsTest.java
@@ -1,5 +1,6 @@
 package com.engineerLiberty.techieAssessment;
 
+import com.engineerLiberty.techieAssessment.zrandomcodes.TimeInWords;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
